+++ examples/specimens/package.json
@@ -23,7 +23,7 @@
     "@loadable/component": "^5.15.2",
     "@mdx-js/react": "^2.1.5",
     "@theme-ui/mdx": "~0.15.3",
-    "@theme-ui/presets": "0.15.3",
+    "@theme-ui/preset-system": "0.15.3",
     "gatsby": "^5.0.0",
     "gatsby-plugin-catch-links": "^5.0.0",
     "gatsby-plugin-manifest": "^5.0.0",
