+++ plop-templates/theme-core-parent/package.json.hbs
@@ -22,7 +22,7 @@
   "dependencies": {
     "@lekoarts/gatsby-theme-{{kebabCase name}}-core": "^1.0.0",
     "@theme-ui/mdx": "~0.15.0",
-    "@theme-ui/presets": "~0.15.0",
+    "@theme-ui/preset-tailwind": "~0.15.0",
     "gatsby-plugin-theme-ui": "~0.15.0",
     "theme-ui": "~0.15.0"
   },
