+++ plop-templates/theme-core-parent/theme-ui.js.hbs
@@ -1,5 +1,5 @@
 import { merge } from "theme-ui"
-import { tailwind } from "@theme-ui/presets"
+import tailwind from "@theme-ui/preset-tailwind"
 
 const theme = merge(tailwind, {
   config: {
