+++ plop-templates/theme/package.json.hbs
@@ -21,7 +21,7 @@
   },
   "dependencies": {
     "@theme-ui/mdx": "~0.15.0",
-    "@theme-ui/presets": "~0.15.0",
+    "@theme-ui/preset-tailwind": "~0.15.0",
     "gatsby-plugin-theme-ui": "~0.15.0",
     "theme-ui": "~0.15.0"
   },
