+++ renovate.json5
@@ -194,6 +194,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -230,6 +232,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -262,6 +266,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -295,6 +301,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -327,6 +335,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -360,6 +370,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -392,6 +404,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -425,6 +439,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -457,6 +473,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -490,6 +508,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -522,6 +542,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -555,6 +577,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -587,6 +611,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -620,6 +646,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -652,6 +680,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -685,6 +715,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -717,6 +749,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -750,6 +784,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -782,6 +818,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -815,6 +853,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -847,6 +887,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -880,6 +922,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -912,6 +956,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -945,6 +991,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -977,6 +1025,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -1010,6 +1060,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
@@ -1042,6 +1094,8 @@
         "@theme-ui/color",
         "@theme-ui/mdx",
         "@theme-ui/presets",
+        "@theme-ui/preset-system",
+        "@theme-ui/preset-tailwind",
         "@theme-ui/prism",
         "@theme-ui/sidenav",
         "gatsby-plugin-theme-ui",
