+++ themes/gatsby-theme-cara/package.json
@@ -25,7 +25,7 @@
     "@react-spring/parallax": "^9.5.5",
     "@react-spring/web": "^9.5.5",
     "@theme-ui/mdx": "~0.15.3",
-    "@theme-ui/presets": "0.15.3",
+    "@theme-ui/preset-tailwind": "0.15.3",
     "gatsby-plugin-mdx": "^5.0.0",
     "gatsby-plugin-theme-ui": "~0.15.3",
     "gatsby-source-filesystem": "^5.0.0",
