+++ themes/gatsby-theme-cara/src/gatsby-plugin-theme-ui/index.js
@@ -1,5 +1,5 @@
 import { merge } from "theme-ui"
-import { tailwind } from "@theme-ui/presets"
+import tailwind from "@theme-ui/preset-tailwind"
 
 const theme = merge(tailwind, {
   config: {
