+++ themes/gatsby-theme-graphql-playground/package.json
@@ -25,7 +25,7 @@
     "@mapbox/rehype-prism": "^0.8.0",
     "@mdx-js/react": "^2.1.5",
     "@theme-ui/mdx": "~0.15.3",
-    "@theme-ui/presets": "0.15.3",
+    "@theme-ui/preset-tailwind": "0.15.3",
     "@theme-ui/sidenav": "~0.15.3",
     "gatsby-plugin-mdx": "^5.0.0",
     "gatsby-plugin-theme-ui": "~0.15.3",
