+++ themes/gatsby-theme-jodie/package.json
@@ -23,7 +23,7 @@
     "@emotion/react": "^11.10.5",
     "@lekoarts/gatsby-theme-jodie-core": "^5.0.0",
     "@theme-ui/mdx": "~0.15.3",
-    "@theme-ui/presets": "0.15.3",
+    "@theme-ui/preset-tailwind": "0.15.3",
     "gatsby-plugin-image": "^3.0.0",
     "gatsby-plugin-theme-ui": "~0.15.3",
     "polished": "^4.2.2",
