+++ themes/gatsby-theme-minimal-blog/package.json
@@ -26,7 +26,7 @@
     "@mdx-js/react": "^2.1.5",
     "@theme-ui/color": "~0.15.3",
     "@theme-ui/mdx": "~0.15.3",
-    "@theme-ui/presets": "0.15.3",
+    "@theme-ui/preset-tailwind": "0.15.3",
     "gatsby-plugin-catch-links": "^5.0.0",
     "gatsby-plugin-theme-ui": "~0.15.3",
     "prism-react-renderer": "^1.3.5",
