+++ themes/gatsby-theme-status-dashboard/package.json
@@ -22,7 +22,7 @@
   "dependencies": {
     "@emotion/react": "^11.10.5",
     "@theme-ui/mdx": "~0.15.3",
-    "@theme-ui/presets": "0.15.3",
+    "@theme-ui/preset-tailwind": "0.15.3",
     "gatsby-plugin-theme-ui": "~0.15.3",
     "gatsby-source-circleci": "^2.0.0",
     "gatsby-source-netlify": "^2.0.0",
