+++ www/package.json
@@ -17,7 +17,7 @@
     "@mdx-js/react": "^2.1.5",
     "@react-three/drei": "^9.40.0",
     "@theme-ui/mdx": "~0.15.3",
-    "@theme-ui/presets": "0.15.3",
+    "@theme-ui/preset-tailwind": "0.15.3",
     "gatsby": "^5.0.0",
     "gatsby-plugin-catch-links": "^5.0.0",
     "gatsby-plugin-google-gtag": "^5.0.0",
